refactor: reorganize component structure and improve imports
- Create logical folder structure for components:
  - ui/common for UI utility components (ChartError, ChartSkeleton, etc.)
  - ui/upload for upload-related components (UploadPage, MultiFileUpload)
  - visualizations/ with subdirectories for each visualization type
  - layout/ for structural components (Header, Footer)

- Improve file organization:
  - Move visualization components to their dedicated folders
  - Group related components together (distribution, sentiment, etc.)

- Enhance import patterns:
  - Replace relative imports with @/ alias pattern
  - Update all import paths throughout the codebase
  - Fix broken imports after reorganization

- Component improvements:
  - Rename SentimentChart_v2 to SentimentChart
  - Make Logo component unselectable and link to home page

- Code cleanup:
  - Remove duplicate WordCloud directory
  - Delete unnecessary redirect files
  - Ensure consistent naming conventions
