--- src/components/WordCloud.tsx
@@ -1,3 +1,0 @@
-// Re-export the refactored WordCloud component
-import WordCloud from './WordCloud/index';
-export default WordCloud;
