+++ src/components/ui/common/ChartError.tsx
