+++ src/components/visualizations/distribution/ChartModal.tsx
