+++ src/components/visualizations/wordcloud/store.ts
